Refactor : 테스트 패키지 구조 수정, 설정 파일 수정
