+++ src/main/resources/application-secret.yml.default
@@ -3,5 +3,6 @@ custom:
     name: 'DATABASE NAME'
     username: 'DATABASE USERNAME'
     password: 'DATABASE PASSWORD'
-   jwt:
+  jwt:
      secretKey: "JWT SECRET_KEY"
+     accessTime :
