+++ src/main/resources/application-test.yml
@@ -4,3 +4,10 @@ spring:
     url: jdbc:h2:mem:test;
     username: sa
     password:
+  jpa:
+    hibernate:
+      ddl-auto: create
+custom:
+    jwt:
+      secretKey: "JWT SECRET_KEY/JWT SECRET_KEY/JWT SECRET_KEY/JWT SECRET_KEY"
+      accessTime : "#{60 * 60 * 24}"
