--- src/test/java/com/wanted/WantedApplicationTests.java
@@ -1,13 +1,0 @@
-package com.wanted;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class WantedApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
